refactor: introduce base model
